refactor(web): extract partition-form transformations to dedicated module
Moves data transformation logic from Form.tsx to a new transformations.ts
module:

- buildPayload: converts form values to ConfigModel.Partition
- toFormValues: converts ConfigModel.Partition to initial form values
- useSolvedSizes: calculates min/max sizes via backend solver

These functions use shared helpers from shared/transformations.ts for
filesystem and size configuration, adding partition-specific logic for the
name field (used when reusing existing partitions).

Also consolidates use-solved-sizes.ts into transformations.ts for better
organization.

Reduces Form.tsx from 516 to ~340 lines by separating pure transformation
logic from UI orchestration.
